output error fix

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -11,17 +11,17 @@
  */
 int interpolation(int *a, size_t lo, size_t hi, int x)
 {
-	int p;
+	size_t p;
 
 	if (lo <= hi)
 	{
 		p = lo + (((double)(hi - lo) / (a[hi] - a[lo])) * (x - a[lo]));
 		if (p > hi)
 		{
-			printf("Value checked array[%d] is out of range\n", p);
+			printf("Value checked array[%ld] is out of range\n", p);
 			return (-1);
 		}
-		printf("Value checked array[%d] = [%d]\n", p, a[p]);
+		printf("Value checked array[%ld] = [%d]\n", p, a[p]);
 		if (a[p] == x)
 			return (p);
 		if (a[p] < x)
